browser: set the headers on the page
Trying to

```
Browser.fetch("https://handicap.gouv.fr")
```

is rejected, presumably because of the wrong user-agent being sent,
which by default (Ferrum) includes the `ChromeHeadless` string which I
suspect is being blocked by the website.

Ferrum does some `User-Agent` override[1] when it sets headers, but not
before[2]. This means that we must:

1. set the correct `default_user_agent` on the browser object;
2. set the headers on the page for the override to kick in.

Arguably this might be a bug on their side (setting the browser's
default_user_agent should suffice, not relying on `headers.set` side
effects) but in the meantime this works.

[1]: https://github.com/rubycdp/ferrum/blob/main/lib/ferrum/headers.rb#L71
[2]: https://github.com/rubycdp/ferrum/blob/main/lib/ferrum/headers.rb#L52-L62