gem: enable using debug
ruby/debug is the new debugger on the block except it's maintained by
the core team and is just as powerful as pry or byebug. Since the
initial template is configured for `debug` reintroduce the gem and
make the necessary config tweaks to launch it (namely disabling
multi-threading which is error-prone in local dev).