check: introduce a state machine
The checks and their execution are very much a state-machine: use the
dedicated statesman library which besides a really nice API offers an
ActiveRecord interface to track transitions between states.

This is how it goes: an Audit is a collection of checks which must all
finish at some point regardless of their outcome (i.e you might fail
to declare your HTML lang attribute, but the check *did*
finish/complete).

Checks are then:

1) pending   : the check has been requested
2) ready     : the check can run
3) blocked   : the check cannot run because of unmet requirements (for now)
4) running   : the check is being ran in a job
5) failed    : the check could not complete, even after retries
6) completed : the check has completed, regardless of the result

The nice part? ActiveJob does a lot of it: since it controls the
execution of the job, we can leverage its retry and fail mechanisms
to handle most of the state changes.