jobs: ensure we store the root cause of a Check::RuntimeError
We can use Ruby's `Exception.cause`[1] to set the root cause of a
RuntimeError and leverage the whole API to set the correct information
in a failed transition's metadata.

We also leverage the `json/add/exception`[2] bit to ensure that the
serialization of the error is properly structured rather than the
default `to_s` which yields the error message or the error class if
there is no message. This allows us to get:

{
  "json_class": "SomeErrorClass",
  "m": "the error message",
  "b": "the backtrace"
}

[1]: https://docs.ruby-lang.org/en/3.4/Exception.html#method-i-cause
[2]: https://docs.ruby-lang.org/en/3.4/Exception.html#method-i-as_json