browser: actually use a blocklist
We want to set the blocklist on the page's context rather than the
browser's root context/page, which we never use: we use a dedicated
page (c.f create_page) per check, which is logical and also the
recommended way in the Ferrum docs.

Which is why after a bit of testing turns out nothing was blocked
which means Chrome ended up making lots of requests and probably
saturating the max number of connections on the websocket (which
appears to be ~256 from a quick online search).

This fix should buy us enough time to fine-tune the process and
prevent the daunted Ferrum::DeadBrowserError/Ferrum::TimeoutError,
because a lot less requests should be made and Chrome should survive
much longer.