jobs: add FetchHomePageJob
FetchHomePage's responsibility is to:

1. fetch an audit's home page HTML
2. store it on the audit
3. tell the audit to start some checks.

This has multiple benefits, notably isolating the error-prone
ressource fetching that was previously triggered by checks themselves,
in `RunCheckJob`: using `Check#root_page` was triggering a `Page.new`
which itselfs triggers a `fetch_page_data` which triggers a
`Browser.get` if that fetch hadn't been cached/stored before.

So we're flipping things around by first getting the resource (and
leveraging ActiveJob's API to handle all the network issues) and then
running checks that depend on it. This means the `Check` and `Page`
classes will become a lot less turbulent by being "read-only"
classes: the resource fetching/storing/error-handling is out of the
way when they run.