jobs: bundle the fetching tasks into FetchResourcesJob
It's nice to have both FetchHomePageJob and FetchAccessibilityPageJob
but it requires synchronisation: when FetchHomePageJob finishes
getting the homepage, it will eventually trigger all the jobs without
waiting on the inherently slower ProcessAuditJob:

1. FetchHomePageJob...
2. FetchAccessibilityPageJob...
3. FetchHomePageJob OK -> ProcessAuditJob
4. RunCheckJob(Reachable) OK -> ProcessAuditJob
5. RunCheckJob(FindAccessibilityPage) FAILS (no a11y HTML)
6. All further accessibility page checks are cancelled
7. FetchAccessibilityPageJob OK -> too late.

To work around this we're simply going to bundle up the home page and
find accessibility page business together and just have one big
crawl. You can't get the accessibility page without the home page
anyway, so we're moving away from that distinction and now we'll just:

1. FetchResourcesJob(audit)
2. ProcessAuditJobs will trigger all the jobs.

The whole concept of requirements might eventually disappear because
checks don't depend on each other anymore: they have the resource (or
they don't) and they can act accordingly and independently – that's
probably a good thing.